Added some python 3 compatibility
