+++ cufflinks/utils.py
@@ -2,7 +2,7 @@ import pickle
 
 def merge_dict(d1,d2):
 	d=d2.copy()
-	for k,v in d1.items():
+	for k,v in list(d1.items()):
 		if k not in d:
 			d[k]=v
 		else:
@@ -17,34 +17,34 @@ def pp(el,preString=''):
 	"""
 	tab=' '*4
 	if isinstance(el,dict):
-		keys=el.keys()
+		keys=list(el.keys())
 		keys.sort()
 		for key in keys:
 			val=el[key]
 			if isinstance(val,dict) or isinstance(val,list):
-				print '%s%s :' % (preString,key)
+				print('%s%s :' % (preString,key))
 				pp(val,preString+tab)
 			else:
-				print '%s%s  =  %s' % (preString,key,val)
+				print('%s%s  =  %s' % (preString,key,val))
 	
 	elif isinstance(el,list):
-		print preString+tab+'['
+		print(preString+tab+'[')
 		preString+=tab	
 		for _ in el:
 			if isinstance(_,dict):
-				print preString+tab+'{'
+				print(preString+tab+'{')
 				pp(_,preString+tab*2)
-				print preString+tab+'}'
+				print(preString+tab+'}')
 			elif isinstance(_,list):
-				print preString+tab+'['
+				print(preString+tab+'[')
 				pp(_,preString+tab)
-				print preString+tab+']'
+				print(preString+tab+']')
 			else:
 				pp(_,preString+tab)
 			#print preString+'      '+str('-')*10            
-		print preString	+']'
+		print(preString	+']')
 	else:
-		print preString+str(el)	
+		print(preString+str(el))	
 
 def inverseDict(d):
 	"""
@@ -54,7 +54,7 @@ def inverseDict(d):
 		d : dictionary
 	"""
 	dt={}
-	for k,v in d.items():
+	for k,v in list(d.items()):
 		if type(v) in (list,tuple):
 			for i in v:
 				dt[i]=k
